Cached text of books presplit by sentence
The heaviest part of bootup was loading all the segmented text. While this could
be defered in some way, for now I just wanted to speed it up. Instead of segmenting
the entire text every time (extremly expensive) after the first time doing a full
segmentation we join each sentence in the split text, and save a json array of
each sentence in the text.

Next when doing word lookups, we only have to segment the text if it contains
we are looking for (which is cheap to test for).

This led to a 5x speedup on my machine (10s boot time down to around 1.7s). As well
as no noticable slowdown on generating sentences for flashcards