Move to wails
Still not sure what I want to do going forwards, whether to pick wails,
tauri, stick with electron or go elsewhere. But thanks to the
refactoring I already did to move to tauri, I was actually able to get
this going in only 3 or so hours.

In the present state we still have to launch the node process ourselves,
but bundling it and running it from wails should not prove too difficult
(I hope)