Partial rewrite of UserSettings
Refactored usersettings to be a nested struct for easier readability.
This way also provides strongly typed settings which allow both lsp
completions as well as reativity.

Gluing together the old way of frontend setting componenets with this
new way saving them is messy for the moment, but I will come up with a
cleaner solution later.