Only convert backend <-> frontend card format at load/save
Before we would format things to be ready to send to anki as soon as
they are selected.

Now we define some loading and saving convention. This makes the code
a little more verbose in places but sets up nicely for card management
in the future.

TODO:
* Use originalValues to preselect options in card creation
* Some better expectations around adding new images/definitions
  and how it will play with previous ones

Also first front end unit tests yay!